+++ .gitignore
@@ -5,8 +5,7 @@
 build
 composer.lock
 coverage
-docs
 phpunit.xml
 psalm.xml
-vendor
+/vendor
 .php-cs-fixer.cache
