+++ composer.json
@@ -14,6 +14,11 @@
       "name": "Igniter Labs Team"
     }
   ],
+  "autoload": {
+    "psr-4": {
+      "IgniterLabs\\GiftUp\\": "src/"
+    }
+  },
   "extra": {
     "tastyigniter-extension": {
       "code": "igniterlabs.giftup",
