+++ resources/views/settings/info.blade.php
